bugfix: invalid storage class selected for some size config parameters.
This resulted in wrong values. The most prominent victim was the
directory creation mode, which was set to zero in some cases. For
details, see related blog post:
http://blog.gerhards.net/2009/10/another-note-on-hard-to-find-bugs.html

This replaces the improper bugfix from two commits ago with a proper one.

+++ tools/syslogd.c
@@ -1748,7 +1748,7 @@ static void doexit()
 
 
 /* set the maximum message size */
-static rsRetVal setMaxMsgSize(void __attribute__((unused)) *pVal, int iNewVal)
+static rsRetVal setMaxMsgSize(void __attribute__((unused)) *pVal, long iNewVal)
 {
 	return glbl.SetMaxLine(iNewVal);
 }
